singleton pattern for bankBranch class
Added a 'bankBranchInstance' variable to hold the single instance of bankBranch class. Defined the bankBranch class to manage a constructor that checks if the instance already exists using bankBranchInstance, if no instance exists it creates one and assigns it to bankBranchInstance and if one does exist it returns the existing one. This makes sure theres only one instance of bankBranch and ensures the info is the same across the webpage.